pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "lyrebird"
version = "0.1.0"
description = "Public-to-private GitHub issue mirroring"
requires-python = ">=3.10"
dependencies = [
    "PyGithub>=2.1.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
]

[tool.setuptools.packages.find]
include = ["lyrebird*"]

[project.scripts]
lyrebird = "lyrebird.cli:main"
